requirements.txt
google-genai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
python-dotenv>=1.0.0
umap-learn>=0.5.0
hdbscan>=0.8.0
tqdm>=4.65.0